W4D1/Q1/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.3
langchain==0.2.1
pinecone-client==3.2.2
python-dotenv==1.0.1
pydantic==2.7.1
PyPDF2==3.0.1
requests==2.32.3
tqdm==4.66.4
langchain-openai==0.1.5 